Updated VideoRehabService to use BaseWebRTCService
